Fix files tests
They made too many assumptions about how attachments work, which were
invalidated by refactors of Files. We no longer return links with
`chatkit_link`, we always return direct links. The logic in the SDK is
sound, but the test wasn't using it, and rather assuming it knew what
type of fetch was needed.