Remove tests for event emitting on initial connect
These tests are actually in error - we do not emit events describing the
initial state, and `disconnect` clears the internal state.

The cases these tests were intended to cover are actually those where
the client has become unintentionally disconnected, and are tested by
the state replacement tests added to the stores.