Remove Jacoco code coverage plugin
It was breaking the build for reasons which aren't worth investigating,
as we aren't making use of the reports.

Whenever coverage is added for critical functional, the coverage
percentage seems to slightly decrease (perhaps as a result of
refactoring the functionality for testability), which is unhelpful. Raw
line-by-line test coverage is a vanity metric, not a useful measure of
test quality.