Add parsing membership in InitialState of user's subscription
Make it easier to investigate malformed JSON failures
in UserSubscriptionEventParserSpec (Gson reports line/column info
so it's good to have JSONs in separate files).

Move ReadStateApiType into UserSubscriptionEvent file as having
a separate file for it was giving it too much exposure and was
creating a bit of confusion as a few other files in that dir
are also related to the backend API but are not explicitly named
to convey that. Normally having all BE API responsibilities inside
api package would be good but currently it would require too much
of refactoring.