Improve tests
Improve one test so that it does not wrongly assume particular order
of events and so that it verifies more precisely lastMessageAt values.

Use consistently outer (sealed) class name when referring to
inner subclasses to improve legibility (of dealing with a public
ChatEvent type) and to fix warnings.

Improve another test to not use legacy V2 backend core API as it was
not intended for that test.

Improve another test to clearly call v2 legacy and have expected
deprecated API usage not triggering warnings.