Fix generating AddedToRoomEvent on reconnect
Read states are returned only for the top 1000 rooms hence for that
edge case when connection is lost and re-established, trying to find
a read state for a new joined room that falls beyond the top 1000 limit
was causing NPE.

The fix aims to be a minimal change that does not escalate optionality
but puts the compiler to work for us. A better solution would be to
have a clear separation between the external, internal and public types.
It would be however a broader change, hence it's something to consider
later.

Simplify reporting language of UserSubscriptionEventParserSpec.kt
using the present tense.