Fix RC in processing initial state on connect and subsequent events
Fix RC in public facing room list retrieval (introduced one commit ago
with synchronized map replacement of ConcurrentMap).

Add test for the top 1000 joined rooms scenario when read_state_updated
event arrives for a room that originally came without read status.

Improve FutureValue so that calling set more than once is clearly
communicated as an error and so that there's no need for additional
event counters verifying the expected number of events
(the tests that had such counters are simplified).

Improve RoomUpdated events tests for the sending message scenario
to truly not rely on order of the events.

Improve use of FutureValue making it a delegated property in one test.

Upgrade the calls in RoomSpek to not use deprecated versions of the API.