Add CursorStoreReceivingReadStateUpdatedTest
Rename CursorsStore -> CursorStore — consistency with RoomStore,
correct English and sounds better (it's obvious that it stores
multiple cursors, not a single one as that would not make sense).

Rename CursorStoreSpec -> CursorStoreApplyingNewInitialStateTest
for specific name.

Rename and repackage RoomStoreReceivingNewInitialStatesSpek ->
RoomStoreApplyingNewInitialStatesTest for consistency.

Improve assertSingletonListWithElementType fun to be more consistent
and Kotlin-idiomatic (consulted with Mike).