Remove dependency on JVM 7 extension of kotlin-stdlib
We want chatkit-core to be as pure Kotlin as it is reasonably possible
at the current state of the project.
More insight:
https://medium.com/@mbonnin/the-different-kotlin-stdlibs-explained-83d7c6bf293
https://stackoverflow.com/questions/51858596/which-standard-library-to-use-in-kotlin

The article actually suggests also removing it for Android specific
stuff but as that's automatically added by AS when creating a new
Android project, I think it's better to trust AS.
The downside mentioned in the article is just 3k bloat which will be
there most likely either way for any Android Kotlin app,
so not a biggie.