Fix SdkInfo version to not be always 1.0
Adjust ChatkitDependencies for v2 work

It has been agreed with Mike to no longer accept custom OkHttp instance.
Likewise it has been agreed to require Android Context as it will be
needed for date/time formatting respecting device settings
in ViewModels.

Default logger is empty implementation (no logging).

Simpler and not misleading setup. For example not accepting
optional PlatformDependencies when we override separately them...

Removed stuff that won't be used in the initial iterations on v2:
AppHookEmitter and PushNotificationsFactory.

There's a working open class ChatkitDependencies (chatkit-core)
with NullMediaTypeResolver as opposed to just an interface.
Setting SDK_VERSION also at chatkit-core level, not ideal now as
required redundancy of the definition to what is used for publishing.
It's not a biggie though as it's only used for analytics.
Bumping up a ver no should be done as find-replace-all using the old
version. There is also a test with explicit test data for it now.

Put precise values in SdkInfo's. Reuse it between the core and android.
Fixed the legacy values for the tests.

Cleaner naming in the test code.