Create our StoreSubscriber
Thanks to it a single dependency will be needed for the
classes that subscribe to the store as observers will be notified
with the current state object (as opposed to having to get
a separate GetState dependency to retrieve the current state while
being just notified that something has changed, i.e. () -> Unit).

Additionally StoreSubscription has a clearly named unsubscribe
function, as opposed just to subscribe fun return type of () -> Unit.

Also with having a more concrete type there's no need for named
qualifier which was added for clarity and clash avoidance on:
(() -> Unit) -> (() -> Unit).